- Added new Tool Shears - Added Bismuth Bronze - Refactoring&Tweaking of deposit occurences - Refractoring of Tree Sapling drop rate - Many Game Balance Tweaks - Added some 32x32 Textures for vanilla items - Reworked gfx of Stone and Bronze Tools - Bugfixes

+++ src/main/java/at/tyron/vintagecraft/gui/GuiStove.java
@@ -70,7 +70,7 @@ public class GuiStove extends GuiContainer {
 	
 	
     private int smeltprogress(int multiplicator) {
-        int orecookingtime = stove.oreCookingTime;
+        int orecookingtime = (int) stove.oreCookingTime;
         int maxcookingtime = stove.maxCookingTime();
         
         return maxcookingtime != 0 && orecookingtime != 0 ? orecookingtime * multiplicator / maxcookingtime : 0;
